Added YouTube Assets and get_eclipse_images
Added the YouTube Assets folder to keep all of the assets and files needed to make the parts that make up the YouTube API call.

Also added the get_eclipse_images function as the adding of image files for the movie will become far more complex as we start to incoporate participants' photos.