Added escsp_read_eclipse_csv and escsp_get_eclipse_time_trio
Made the function escsp_read_eclipse_csv in the escsp module because it was repeated in multiple functions and not working properly.  It will be used enought that it will be handy to hae as a separate function.
Made the function escsp_get_eclipse_time_trio in the escsp module so that I didn't have to recreate the eclipse times on multiple days as it is needed in multiple functions.