read_sd_cards.py
Created the read_sd_cards.py script to read an AudioMoth SD card and automatically file it in the ES data server.  It is currently a serial program. Considering the time it takes to read a file, I need to parallelize it.

I also made the escsp_sn2esid function in the escsp.py module to return the ESID number based on a CONFIG.TXT file and a pandas data frame created from the AudioMoth Master spreadsheet.