Fixing type errors in programs looking for files.
Added documentation to AM_timestamp_set function and put all debugging print statements into if verbose: statements.

Added a script to search all folders in a top level directory.  if  the AM_timestamp_set function reports true, it will split every .WAV file into 1 minute chunks and then save them into a defined output folder.

Changed "type" parameter to "file_type" in several codes so that it wouldn't confuse with the python function and type.

The filename_2_datetime function had an error when it came to the use of the variable "type" which caused errors intermittently when the reports code was searching to find if a folder contained data files within a certain date.

Other functions contained the same logic error and were corrected.

Created a test_reports_1.py script to test the Reports_1 function rather than having it run on an entire folder.