YouTube Updates
Changed the file structure so that there is a YouTube_Queue and a  YouTube_Uploaded file.

Created the function get_youtube_link to search the output of the YouTube API call for link information.

Changed subprocess.call() -> subprocess.run() which not only allows me to get stdout and stderr messages, but is also now considered best practice to use.