Added Birdnet Addressed Conflicted escsp.py
Added the script #ESCSP_Analysis_01.py to the directory to more easily access the escsp.py module.

During that process, I noticed that there were functions missing from  escsp.py.  Dropbox had generated a conflicted copy of the ESCSP folder and I had not noticed.  I ran a diff on the two different versions of escsp.py to ensure that the version not listed as conflicted had all functions in it.